+++ src/sampen_cpu.cu
@@ -2,19 +2,17 @@
 #include <stdlib.h>
 #include <math.h>
 
-#include "./sampen_cpu.h"
 
 
-
-double chebyshev_distance(double* A, double* B, unsigned int length) {
-    double d = 0;
+float chebyshev_distance(float* A, float* B, unsigned int length) {
+    float d = 0;
     for(unsigned int i = 0; i < length; i++) {
         d = max(abs(A[i] - B[i]), d);
     }
     return d;
 }
 
-double sampen_cpu(double* data, unsigned int length, unsigned int m, double r) {
+float sampen_cpu(float* data, unsigned int length, unsigned int m, float r) {
     if(m < 1) { return 0; }
     unsigned long A = 0, B = 0;
 
@@ -32,6 +30,6 @@ double sampen_cpu(double* data, unsigned int length, unsigned int m, double r) {
     }
 
     printf("A=%ld, B=%ld\n", A, B);
-    printf("A/B=%16.16lf\n", (double)A / (double)B);
-    return -log((double)A / (double)B);
+    printf("A/B=%16.16f\n", (float)A / (float)B);
+    return -log((float)A / (float)B);
 }
